Rewritten a code. Apple.h Snake.h and DeadPool.h was deleted and created new header file to describe the three classes Created Apple class and describe his mehtods Fixed Snake move trable

+++ Snake.cpp
@@ -1,5 +1,7 @@
-#include"Snake.h"
-#include<vector>
+#include"Stack_of_class_for_Game.h"
+
+
+
 
     void Snake::Snake_lenght_resize() {
 
@@ -8,7 +10,7 @@
     }
 
 
-   Snake:: Snake(double snake_spead) :snake_spead(snake_spead) { snake_lenght=1; }
+   Snake:: Snake(int snake_spead) :snake_spead(snake_spead) { snake_lenght=1; }
 
    Snake:: ~Snake() {}
